+++ src/molecules/ChipLarge.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { StyleSheet, Text, View, ViewStyle } from 'react-native';
+import { MaybeTouchable } from 'src/components/MaybeTouchable';
 import { colors, isLight } from 'src/utils/colors';
-import { Icon } from 'src/utils/const';
+import { Icon, onIOS } from 'src/utils/const';
 import { registeredTextStyle } from 'src/utils/textStyles';
 
 const styles = StyleSheet.create({
@@ -22,7 +23,7 @@ const styles = StyleSheet.create({
   TextDark: registeredTextStyle('titleBold', colors.uma, 'chip-text-dark'),
 });
 
-type ChipLargeProps = {
+export type ChipLargeProps = {
   label: string;
   icon?: string;
   color: string;
@@ -31,27 +32,38 @@ type ChipLargeProps = {
    * if there are many chips on screen)
    */
   colorIsLight?: boolean;
+  onPress?: () => any;
 };
 
+const renderChip = (color: string, darkText: boolean, label: string, icon?: string) => (
+  <View style={[styles.Wrapper, { backgroundColor: color }]}>
+    {icon ? (
+      <View style={styles.IconAndText}>
+        <Icon
+          style={styles.Icon}
+          name={icon}
+          size={20}
+          color={darkText ? colors.uma : colors.ulisse}
+        />
+        <Text style={darkText ? styles.TextDark : styles.Text}>{label.toUpperCase()}</Text>
+      </View>
+    ) : (
+      <Text style={darkText ? styles.TextDark : styles.Text}>{label.toUpperCase()}</Text>
+    )}
+  </View>
+);
+
 const ChipLargeUnmemoized = (props: ChipLargeProps) => {
-  const { color, colorIsLight } = props;
+  const { color, colorIsLight, icon, label, onPress } = props;
   const darkText = colorIsLight ?? isLight(color);
-  return (
-    <View style={[styles.Wrapper, { backgroundColor: color }]}>
-      {props.icon ? (
-        <View style={styles.IconAndText}>
-          <Icon
-            style={styles.Icon}
-            name={props.icon}
-            size={20}
-            color={darkText ? colors.uma : colors.ulisse}
-          />
-          <Text style={darkText ? styles.TextDark : styles.Text}>{props.label.toUpperCase()}</Text>
-        </View>
-      ) : (
-        <Text style={darkText ? styles.TextDark : styles.Text}>{props.label.toUpperCase()}</Text>
-      )}
+  return onPress ? (
+    <View style={{ flex: 1, alignItems: 'center' }}>
+      <MaybeTouchable onPress={onPress} exactHeight={22} borderRadius={11}>
+        {renderChip(color, darkText, label, icon)}
+      </MaybeTouchable>
     </View>
+  ) : (
+    renderChip(color, darkText, label, icon)
   );
 };